feat : register, login, pagination dll

+++ app/Http/Controllers/IdeaController.php
@@ -17,18 +17,32 @@ class IdeaController extends Controller
         $validate = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
+
+        // menyimpan id user yang sedang login
+        $validate['user_id'] = auth()->id();
+
         // menggunakan method statis pada sebuah model yaitu create
          Idea::create($validate);
         return redirect()->route('dashboard')->with('success', 'Idea berhasil dibuat!');
     }
     public function destroy(Idea $idea)
     {
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         // menggunakan pengikatan model rute yang lebih sederhana
         $idea->delete();
         return redirect()->route('dashboard')->with('success', 'Idea berhasil dihapus!');
     }
     public function edit(Idea $idea)
     {
+        // jika user yang sedang login bukan user yang membuat idea maka tidak bisa edit
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         // memuat variabel editing dan selanjutnya diteruskan ke view
         $editing = true;
         // kita tidak perlu membuat array assosiatif, cukup menggunakan compact saja
@@ -36,6 +50,10 @@ class IdeaController extends Controller
     }
     public function update(Idea $idea)
     {
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         $validate = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
